Fix broken link in CONTRIBUTING.md
The link to the Apache-2 license was broken, pointing to a non-existant
LICENSE.md file. The top level `LICENSE` file has no extension and linking to
it makes the link a download link in the rendered html, rather than just
displaying the content. The link now point to `LICENSES/Apache-2.0.txt` which
works both for markdown and the published documentation on atomvm.net.

Signed-off-by: Winford <winford@object.stream>