Fix memory leaks in pioc_w network driver
Fixes a leak in `get_default_device_name()`, if the driver if stopped the memory
holding the device name is never free'd. This could be especially problematic if
the driver is stopped and restarted multiple times, as may be the case with a
battery powered device.

Fixes a potential memory leak of `psk` used for the ap if the given `psk` is
invalid due to not meeting the minimum size requirement.

Signed-off-by: Winford <winford@object.stream>