Add `internal_error` atom to defaultatoms.def
Adds the `internal_error` atom from OTP for returning errors for "thing that shouldn't happen".
There are rare occasions where user inputs may be valid, but an internal operation fails for some
reason (potentially an internal bug to the VM, not the users application code). For example, this
may be an invalid internal state when setting the direction for a gpio pin. If the users inputs are
valid `internal_error` should be returned rather than `badarg`, which could mislead application
developers into chasing a bug in their application that isn't there.

Signed-off-by: Winford <winford@object.stream>