Make build-and-test-on-freebsd fail if steps to not complete successfully
This workflow is a different from most in that the entire build and test is done in a single script
in a FreeBSB QEMU constainer on an Ubuntu host. Because it is one large shell script, with no
separate "jobs", exits needed to be added to ensure the test fails if any important steps fail to
complete successfully.

Adds missing rebar3 installation required to complete the tests build.

Signed-off-by: Winford <winford@object.stream>